+++ include/nbsim/bounding_box.h
@@ -8,12 +8,12 @@
  * A cuboid bounding box
  */
 struct BoundingBox {
-    Vec3 center;  // Center of this bounding box
-    float width;  // Width from one side to another
-    BoundingBox(float width) : center{Vec3{0, 0, 0}}, width{width} {}
-    BoundingBox(float x, float y, float z, float width)
+    Vec3 center;   // Center of this bounding box
+    double width;  // Width from one side to another
+    BoundingBox(double width) : center{Vec3{0, 0, 0}}, width{width} {}
+    BoundingBox(double x, double y, double z, double width)
         : center{Vec3{x, y, z}}, width{width} {}
-    BoundingBox(const Vec3& center, float width)
+    BoundingBox(const Vec3& center, double width)
         : center{center}, width{width} {}
 };
