+++ src/object.cpp
@@ -2,7 +2,7 @@
 
 Object::Object() : mass{0}, position{Vec3{0, 0, 0}} {}
 
-Object::Object(float mass, const Vec3& position)
+Object::Object(double mass, const Vec3& position)
     : mass{mass}, position{position} {}
 
 bool Object::operator==(const Object& other) const {
